Refactor build_network to have its functions be more general
